add erc contract creation tx analyze logic

+++ browser-service/src/main/java/com/platon/browser/v0151/bean/Erc20ContractId.java
@@ -3,7 +3,7 @@ package com.platon.browser.v0151.bean;
 import lombok.Data;
 import org.apache.commons.lang3.StringUtils;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 /**
  * Erc20合约标识
@@ -12,8 +12,8 @@ import java.math.BigInteger;
 public class Erc20ContractId {
     private String name;
     private String symbol;
-    private BigInteger decimal;
-    private BigInteger totalSupply;
+    private Integer decimal;
+    private BigDecimal totalSupply;
     public boolean isSupportErc20(){
         return !StringUtils.isBlank(name) && !StringUtils.isBlank(symbol) && decimal != null && totalSupply != null;
     }
